+++ .github/workflows/build.yml
@@ -43,6 +43,8 @@ jobs:
       - name: Install dependencies
         run: yarn
       - name: Build
+        run: yarn build:agent
+      - name: Build
         run: yarn build:message-kit
       - name: Test parsing
         run: yarn test:parsing
