+++ README.md
@@ -9,6 +9,7 @@ This is the official repository for [MessageKit](https://message-kit.org/). Powe
 - [`message-kit`](/packages/message-kit): A kit for quickly building messaging apps
 - [`create-message-kit`](/packages/create-message-kit): A CLI for creating new apps easily
 - [`framekit`](/packages/framekit): A suite for backend frames
+- ['agentsdk](/packages/agent/): A secure agent sdk
 - [`docs`](/packages/docs): Documentation for MessageKit
 
 ## Contributing
