+++ packages/agent/README.md
@@ -1,0 +1,13 @@
+# @xmtp/secure
+
+A secure XMTP client implementation with built-in encryption and key management.
+
+## Features
+
+- Secure key management
+- Support for both v2 and v3 XMTP clients
+- Built-in encryption
+- Support for multiple content types
+- Group messaging capabilities
+
+## Installation
