+++ packages/agent/package.json
@@ -1,0 +1,58 @@
+{
+  "name": "xmtp-agent",
+  "version": "1.0.1",
+  "license": "MIT",
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "require": "./dist/index.cjs",
+      "import": "./dist/index.js"
+    }
+  },
+  "main": "dist/index.cjs",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist/**/*"
+  ],
+  "scripts": {
+    "build": "rollup -c",
+    "build:watch": "yarn build -w",
+    "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
+    "publish": "npm publish"
+  },
+  "dependencies": {
+    "@changesets/changelog-git": "^0.2.0",
+    "@changesets/cli": "^2.27.5",
+    "@xmtp/content-type-primitives": "^1.0.2",
+    "@xmtp/content-type-reaction": "^1.1.10-beta.1",
+    "@xmtp/content-type-read-receipt": "^2.0.0",
+    "@xmtp/content-type-remote-attachment": "1.1.11",
+    "@xmtp/content-type-reply": "^1.1.11",
+    "@xmtp/content-type-text": "^1.0.0",
+    "@xmtp/grpc-api-client": "^0.2.8",
+    "@xmtp/node-sdk": "^0.0.31",
+    "@xmtp/xmtp-js": "^13.0.3",
+    "dotenv": "^16.4.5",
+    "typescript": "^5.4.5",
+    "viem": "^2.16.3"
+  },
+  "devDependencies": {
+    "@rollup/plugin-typescript": "^11.1.6",
+    "@types/node": "^20.14.2",
+    "prettier": "^3.3.1",
+    "rollup": "^4.18.0",
+    "rollup-plugin-dts": "^6.1.1",
+    "ts-node": "^10.9.2",
+    "turbo": "^2.2.3"
+  },
+  "packageManager": "yarn@4.5.1",
+  "engines": {
+    "node": ">=20"
+  },
+  "publishConfig": {
+    "access": "public",
+    "registry": "https://registry.npmjs.org/"
+  }
+}
