+++ packages/agent/rollup.config.js
@@ -15,13 +15,9 @@ const external = [
   "@xmtp/proto",
   "@xmtp/node-bindings",
   "@xmtp/grpc-api-client",
-  "@coinbase/coinbase-sdk",
-  "@coinbase/cbpay-js",
   "@xmtp/content-type-read-receipt",
-  "@circle-fin/developer-controlled-wallets",
   "cross-fetch",
   "path",
-  "xmtp-agent",
   "crypto",
   "viem",
   "dotenv",
