+++ packages/agent/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./lib/xmtp.js";
+export * from "./lib/types.js";
+export * from "./content-types/agent-message.js";
