+++ packages/agent/tests/V2.test.ts
@@ -1,9 +1,9 @@
 import { describe, test, expect, vi, beforeEach } from "vitest";
-import { MessageKit } from "../src/lib/core";
+import { MessageKit } from "../../message-kit/src/lib/core";
 import { ContentTypeText } from "@xmtp/content-type-text";
 import { ContentTypeReply } from "@xmtp/content-type-reply";
 import { ContentTypeReaction } from "@xmtp/content-type-reaction";
-import { getMocks } from "./utils";
+import { getMocks } from "../../message-kit/tests/utils";
 
 describe("Context Message Tests", () => {
   // Mock conversation data
