+++ packages/agent/tests/V3.test.ts
@@ -1,7 +1,7 @@
 import { describe, test, expect, vi, beforeEach } from "vitest";
-import { MessageKit } from "../src/lib/core";
+import { MessageKit } from "../../message-kit/src/lib/core";
 import { ContentTypeText } from "@xmtp/content-type-text";
-import { getMocks } from "./utils";
+import { getMocks } from "../../message-kit/tests/utils";
 
 describe("Context Message Tests", () => {
   const {
