+++ packages/create-message-kit/index.js
@@ -7,7 +7,7 @@ import { default as fs } from "fs-extra";
 import { isCancel } from "@clack/prompts";
 import { detect } from "detect-package-manager";
 import pc from "picocolors";
-const defVersion = "1.2.33";
+const defVersion = "patch";
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
 // Read package.json to get the version
