+++ packages/docs/pages/changelog.mdx
@@ -435,7 +435,7 @@ Overall, these changes aim to enhance the functionality of GPT-based agents, imp
 - Replaced one-to-one with Agent example
 - Documented `skills`
 - Renamed examples as templates
-- How to use a second `xmtpClient` client
+- How to use a second `createClient` client
 - Optional parameters concepts/structure#optional-parameters
 - Added changelog
