+++ packages/framekit/src/components/Chat.tsx
@@ -162,8 +162,8 @@ function Chat({ user }: { user: UserInfo }) {
 
   const initXmtp = async (wallet: any) => {
     try {
-      const xmtpClient = await V2Client.create(wallet, { env: "production" });
-      setXmtp(xmtpClient);
+      const createClient = await V2Client.create(wallet, { env: "production" });
+      setXmtp(createClient);
       setIsLoading(false);
     } catch (error) {
       console.error("Error initializing XMTP:", error);
