+++ packages/message-kit/src/helpers/types.ts
@@ -1,37 +1,6 @@
 import { Context } from "../lib/core.js";
-import { ClientOptions, GroupMember } from "@xmtp/node-sdk";
-import { ContentTypeId } from "@xmtp/content-type-primitives";
+import { xmtpConfig } from "xmtp-agent";
 
-export type MessageAbstracted = {
-  id: string; // Unique identifier for the message
-  sent: Date; // Date when the message was sent
-  content: {
-    text?: string | undefined; // Text content of the message
-    reply?: string | undefined; // Reply content if the message is a reply
-    previousMsg?: string | undefined; // Reference to the previous message
-    react?: string | undefined; // Reaction content if the message is a reaction
-    content?: any | undefined; // Any other content
-    params?: any | undefined; // Parameters for the message
-    reference?: string | undefined; // Reference ID for the message
-    skill?: string | undefined; // Skill associated with the message
-  };
-  version: "v2" | "v3"; // Version of the message
-  sender: AbstractedMember; // Sender of the message
-  typeId: string; // Type identifier for the message
-};
-export type GroupAbstracted = {
-  id: string;
-  sync: () => Promise<void>;
-  addMembers: (addresses: string[]) => Promise<void>;
-  addMembersByInboxId: (inboxIds: string[]) => Promise<void>;
-  send: (content: string, contentType?: ContentTypeId) => Promise<string>;
-  isAdmin: (inboxId: string) => boolean;
-  isSuperAdmin: (inboxId: string) => boolean;
-  admins: string[];
-  superAdmins: string[];
-  createdAt: Date;
-  members: GroupMember[];
-};
 export type SkillResponse = {
   code: number;
   message: string;
@@ -66,7 +35,7 @@ export interface AgentWallet {
 }
 export type AgentConfig = {
   // client options from XMTP client
-  client?: ClientOptions;
+  client?: xmtpConfig;
   // private key to be used for the client, if not, default from env
   privateKey?: string;
   // if true, the init log message with messagekit logo and stuff will be hidden
@@ -128,15 +97,6 @@ export interface Skill {
   description: string;
   examples: string[];
   params?: Record<string, SkillParamConfig>;
-}
-
-export interface AbstractedMember {
-  inboxId: string;
-  address: string;
-  accountAddresses: string[];
-  installationIds?: string[];
-  username?: string;
-  ensDomain?: string;
 }
 
 export type MetadataValue = string | number | boolean;
