+++ packages/message-kit/src/index.ts
@@ -1,5 +1,3 @@
-export * from "./xmtp/stream.js";
-export * from "./xmtp/client.js";
 export * from "./lib/core.js";
 export * from "./lib/skills.js";
 export * from "./helpers/types.js";
@@ -7,6 +5,5 @@ export * from "./plugins/gpt.js";
 export * from "./plugins/resolver.js";
 export * from "./plugins/framekit.js";
 export * from "./skills/concierge.js";
-export * from "./plugins/xmtp.js";
 export { Client as V2Client } from "@xmtp/xmtp-js";
 export { Client as V3Client } from "@xmtp/node-sdk";
