+++ packages/message-kit/src/lib/core.ts
@@ -9,7 +9,7 @@ import {
   Conversation as V2Conversation,
 } from "@xmtp/xmtp-js";
 import { GroupMember } from "@xmtp/node-sdk";
-import type { ContentTypeId } from "@xmtp/content-type-primitives";
+import { ContentTypeId } from "@xmtp/content-type-primitives";
 import { ContentTypeText } from "@xmtp/content-type-text";
 import { ContentTypeReply, type Reply } from "@xmtp/content-type-reply";
 import {
@@ -24,36 +24,81 @@ import {
   type Attachment,
 } from "@xmtp/content-type-remote-attachment";
 
-import { chatMemory, defaultSystemPrompt } from "../plugins/gpt.js";
+import { agentReply, chatMemory, defaultSystemPrompt } from "../plugins/gpt.js";
 import { getUserInfo, userInfoCache } from "../plugins/resolver.js";
-import { logMessage, readFile } from "../helpers/utils.js";
+import { logInitMessage, logMessage, readFile } from "../helpers/utils.js";
 import {
   MessageAbstracted,
   GroupAbstracted,
-  Agent,
-  SkillResponse,
   AbstractedMember,
-} from "../helpers/types.js";
+  AgentMessage,
+  ContentTypeAgentMessage,
+} from "xmtp-agent";
+
+import { Agent, SkillResponse } from "../helpers/types.js";
+
 import {
   type ReadReceipt,
   ContentTypeReadReceipt,
 } from "@xmtp/content-type-read-receipt";
 import { WalletService as CdpWalletService } from "../plugins/cdp.js";
 import { WalletService as CircleWalletService } from "../plugins/circle.js";
-import { executeSkill, parseSkill, findSkill } from "./skills.js";
+import {
+  executeSkill,
+  parseSkill,
+  findSkill,
+  filterMessage,
+} from "./skills.js";
 import { logUserInteraction } from "../helpers/utils.js";
 import path from "path";
 import fetch from "cross-fetch";
 import type { AgentConfig } from "../helpers/types";
-import { XmtpPlugin } from "../plugins/xmtp.js";
-import { AgentMessage } from "../content-types/agent-message.js";
-import { ContentTypeAgentMessage } from "../content-types/agent-message.js";
+import { XmtpPlugin, createClient } from "xmtp-agent";
 import { LocalStorage } from "../plugins/storage.js";
+
+// Add at the top of the file
+export let hasInitialized = false;
+
+export function createAgent(
+  agent: Agent,
+): Agent & { run: () => Promise<void> } {
+  let messageKit: MessageKit | null = null; // Ensure a single instance
+
+  return {
+    ...agent,
+    async run() {
+      if (!messageKit) {
+        // Check if MessageKit is already initialized
+        messageKit = new MessageKit(agent);
+      }
+      await messageKit.run();
+    },
+  };
+}
 
 export const awaitedHandlers = new Map<
   string,
   (text: string) => Promise<boolean | undefined>
 >();
+let hasClientInitialized = false;
+// Create a check function that runs when the module is loaded
+function checkInitialization() {
+  if (!hasInitialized && !hasClientInitialized) {
+    console.warn(`\x1b[33m
+  ⚠️  MessageKit is imported but not running!
+    Make sure to call run(agent) to start processing messages
+
+        Example:
+          import { run } from '@xmtp/message-kit'
+            
+          const agent = {/* Your agent definition here */};
+
+          run(agent)
+  \x1b[0m`);
+  }
+}
+// Run the check after a short delay to allow for initialization
+setTimeout(checkInitialization, 1000);
 
 /* Context Interface */
 export type Context = {
@@ -66,7 +111,6 @@ export type Context = {
   version: "v2" | "v3";
   v2client: V2Client;
   agentConfig?: AgentConfig;
-  agent: Agent;
   walletService: CdpWalletService | CircleWalletService;
   sender?: AbstractedMember;
   awaitingResponse: boolean;
@@ -92,6 +136,7 @@ export type Context = {
   ): Promise<void>;
   awaitedHandler: ((text: string) => Promise<boolean | void>) | undefined;
   xmtp: XmtpPlugin;
+  agent: Agent;
 };
 
 /* Context implementation */
@@ -107,17 +152,11 @@ export class MessageKit implements Context {
   version!: "v2" | "v3";
   v2client!: V2Client;
   members?: AbstractedMember[];
-  admins?: string[];
   agentConfig?: AgentConfig;
-  superAdmins?: string[];
-  agent: Agent = {
-    name: "",
-    description: "",
-    tag: "",
-  };
   walletService!: CdpWalletService | CircleWalletService;
   sender?: AbstractedMember;
   awaitingResponse: boolean = false;
+  agent: Agent;
 
   executeSkill: (text: string) => Promise<SkillResponse | undefined> =
     async () => undefined;
@@ -126,10 +165,37 @@ export class MessageKit implements Context {
   awaitedHandler: ((text: string) => Promise<boolean | void>) | undefined =
     undefined;
 
-  private constructor() {
-    // Empty constructor
+  constructor(agent: Agent) {
+    this.agent = agent;
   }
 
+  async run(): Promise<void> {
+    hasInitialized = true;
+
+    // Initialize the clients
+    const { client, v2client } = await createClient(
+      this.handleMessage,
+      this.agent.config?.client,
+    );
+    hasClientInitialized = true;
+    this.client = client;
+    this.v2client = v2client;
+
+    // Store the GPT model in process.env for global access
+    process.env.GPT_MODEL = this.agent.config?.gptModel || "gpt-4o";
+    logInitMessage(client, this.agent);
+
+    const xmtp = new XmtpPlugin(
+      this.client,
+      this.v2client,
+      this.version,
+      this.refConv,
+      this.conversation,
+      this.group,
+      this.message,
+    );
+    this.xmtp = xmtp;
+  }
   static async create(
     conversation: Conversation | V2Conversation,
     message: DecodedMessage | DecodedMessageV2 | undefined,
@@ -138,7 +204,7 @@ export class MessageKit implements Context {
     version?: "v2" | "v3",
   ): Promise<Context | undefined> {
     try {
-      const context = new MessageKit();
+      const context = new MessageKit(agent);
       context.client = client;
       context.v2client = v2client;
       if (conversation instanceof Conversation) {
@@ -165,7 +231,6 @@ export class MessageKit implements Context {
       if (message && message.id) {
         //v2
         const sentAt = "sentAt" in message ? message.sentAt : message.sent;
-        let members: GroupMember[];
         let sender: AbstractedMember | undefined = undefined;
         if (version === "v2") {
           sender = {
@@ -177,7 +242,7 @@ export class MessageKit implements Context {
         } else {
           let group = await (conversation as Conversation);
           await group.sync();
-          members = await group.members();
+          const members = await group.members();
           context.members = members.map((member: GroupMember) => ({
             inboxId: member.inboxId,
             address: member.accountAddresses[0],
@@ -294,8 +359,6 @@ export class MessageKit implements Context {
           version: version ?? "v2",
         };
 
-        //Plugins
-        context.xmtp = new XmtpPlugin(context as unknown as Context);
         //test
         if (context.agentConfig?.walletService === true) {
           if (
@@ -325,6 +388,102 @@ export class MessageKit implements Context {
       return undefined;
     }
   }
+  handleMessage = async (
+    version: "v3" | "v2",
+    message: DecodedMessage | DecodedMessageV2 | undefined,
+  ) => {
+    const conversation = await this.getConversation(message, version);
+    if (message && conversation) {
+      try {
+        const { senderInboxId, kind } = message as DecodedMessage;
+        const senderAddress = (message as DecodedMessageV2).senderAddress;
+        if (
+          //If same address do nothin
+          senderAddress?.toLowerCase() ===
+            this.v2client.address.toLowerCase() ||
+          //If same address do nothin
+          // Filter out membership_change messages
+          (senderInboxId?.toLowerCase() === this.client.inboxId.toLowerCase() &&
+            kind !== "membership_change")
+        ) {
+          return;
+        }
+        const context = await MessageKit.create(
+          conversation,
+          message,
+          { client: this.client, v2client: this.v2client },
+          this.agent,
+          version,
+        );
+        if (!context) {
+          logMessage("No context found" + message);
+          return;
+        }
+        context.xmtp = this.xmtp;
+
+        //Await response
+        const awaitedHandler = awaitedHandlers.get(
+          context.xmtp.getConversationKey(),
+        );
+        if (awaitedHandler) {
+          const messageText =
+            context.message.content.text || context.message.content.reply || "";
+          // Check if the response is from the expected user
+          const expectedUser = context.xmtp.getConversationKey().split(":")[1];
+          const actualSender =
+            version === "v3"
+              ? (message as DecodedMessage).senderInboxId
+              : (message as DecodedMessageV2).senderAddress;
+
+          if (expectedUser?.toLowerCase() === actualSender?.toLowerCase()) {
+            const isValidResponse = await awaitedHandler(messageText);
+            // Only remove the handler if we got a valid response
+            if (isValidResponse) {
+              awaitedHandlers.delete(context.xmtp.getConversationKey());
+            }
+          }
+          return;
+        }
+
+        // Check if the message content triggers a skill
+        const { isMessageValid, customHandler } = await filterMessage(context);
+        if (isMessageValid && customHandler) {
+          const result = await customHandler(context);
+          if (result && "code" in result) {
+            if (result.code === 200) {
+              await context.send(result.message);
+            }
+          }
+        } else if (isMessageValid && this.agent?.onMessage)
+          await this.agent?.onMessage?.(context);
+        else if (isMessageValid && !this.agent?.onMessage)
+          await this.onMessage(context);
+      } catch (e) {
+        console.log(`error`, e);
+      }
+    }
+  };
+  getConversation = async (
+    message: DecodedMessage | DecodedMessageV2 | undefined,
+    version: "v3" | "v2",
+  ): Promise<Conversation | V2Conversation> => {
+    return version === "v3"
+      ? ((await this.client.conversations.getConversationById(
+          (message as DecodedMessage)?.conversationId as string,
+        )) as Conversation)
+      : ((message as DecodedMessageV2)?.conversation as V2Conversation);
+  };
+
+  onMessage = async (context: Context) => {
+    /*Default onMessage function, replaces the prompt file*/
+    const { agent } = context;
+    if (!agent.systemPrompt) {
+      console.log("System prompt is not defined");
+      return;
+    }
+    await agentReply(context);
+  };
+
   async awaitResponse(
     prompt: string,
     validResponses?: string[],
@@ -411,33 +570,18 @@ export class MessageKit implements Context {
     contentType: ContentTypeId = ContentTypeText,
     targetConversation?: V2Conversation,
   ) {
+    await this.xmtp.send(
+      message,
+      contentType,
+      targetConversation ?? this.refConv ?? this.conversation ?? this.group,
+    );
     if (contentType === ContentTypeText && typeof message !== "string") {
       console.error("Message must be a string");
       return;
     }
     let messageString = message as string;
-    if (
-      typeof message === "object" &&
-      message !== undefined &&
-      //@ts-ignore
-      message?.content !== undefined
-    ) {
-      //@ts-ignore
-      messageString = message?.content as string;
-    }
-    const conversation =
-      targetConversation ?? this.refConv ?? this.conversation ?? this.group;
-    if (conversation) {
-      if (this.xmtp.isV2Conversation(conversation)) {
-        await conversation.send(message, {
-          contentType: contentType,
-        });
-      } else if (this.xmtp.isV3Conversation(conversation)) {
-        await conversation.send(message, contentType);
-      }
-      chatMemory.addEntry(this.getMemoryKey(), messageString, "assistant");
-      logMessage("sent:" + messageString);
-    }
+    chatMemory.addEntry(this.getMemoryKey(), messageString, "assistant");
+    logMessage("sent:" + messageString);
   }
   getMemoryKey() {
     return this.xmtp.getConversationKey() + ":" + this.message?.sender?.address;
