+++ templates/ens/src/index.ts
@@ -1,4 +1,4 @@
-import { run, Agent } from "@xmtp/message-kit";
+import { createAgent } from "@xmtp/message-kit";
 import { checkDomain } from "./skills/check.js";
 import { cool } from "./skills/cool.js";
 import { info } from "./skills/info.js";
@@ -6,11 +6,9 @@ import { register } from "./skills/register.js";
 import { renew } from "./skills/renew.js";
 import { pay } from "./skills/pay.js";
 
-export const agent: Agent = {
+export const agent = createAgent({
   name: "Ens Agent",
   tag: "@bot",
   description: "A ens agent with a lot of skills.",
   skills: [checkDomain, cool, info, register, renew, pay],
-};
-
-run(agent);
+}).run();
